requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
anthropic==0.52.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # passlib 1.7.4 is incompatible with bcrypt >= 4.1
pytest==8.3.4
pytest-asyncio==0.25.0
psycopg2-binary==2.9.10
httpx>=0.27.0
